Move GUI events to client event (Follow up to #387)
