+++ src/main/java/de/teamlapen/vampirism/core/ModEventHandler.java
@@ -4,11 +4,9 @@ import de.teamlapen.lib.lib.util.UtilLib;
 import de.teamlapen.lib.lib.util.VersionChecker;
 import de.teamlapen.vampirism.VampirismMod;
 import de.teamlapen.vampirism.api.VampirismAPI;
-import de.teamlapen.vampirism.client.gui.GuiSkills;
 import de.teamlapen.vampirism.config.Balance;
 import de.teamlapen.vampirism.config.Configs;
 import de.teamlapen.vampirism.entity.converted.VampirismEntityRegistry;
-import de.teamlapen.vampirism.entity.factions.FactionPlayerHandler;
 import de.teamlapen.vampirism.modcompat.IntegrationsNotifier;
 import de.teamlapen.vampirism.network.SyncConfigPacket;
 import de.teamlapen.vampirism.util.DaySleepHelper;
@@ -16,17 +14,12 @@ import de.teamlapen.vampirism.util.REFERENCE;
 import de.teamlapen.vampirism.world.ModWorldEventListener;
 import de.teamlapen.vampirism.world.villages.VampirismVillage;
 import de.teamlapen.vampirism.world.villages.VampirismVillageHelper;
-import net.minecraft.client.gui.GuiButton;
-import net.minecraft.client.gui.GuiButtonImage;
-import net.minecraft.client.gui.inventory.GuiInventory;
 import net.minecraft.entity.player.EntityPlayerMP;
-import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.util.text.TextComponentTranslation;
 import net.minecraft.util.text.TextFormatting;
 import net.minecraft.village.Village;
-import net.minecraftforge.client.event.GuiScreenEvent;
 import net.minecraftforge.event.AttachCapabilitiesEvent;
 import net.minecraftforge.event.terraingen.InitMapGenEvent;
 import net.minecraftforge.event.world.WorldEvent;
@@ -39,6 +32,7 @@ import net.minecraftforge.fml.common.gameevent.TickEvent;
 import net.minecraftforge.fml.common.network.FMLNetworkEvent;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
 import java.util.List;
 
 /**
@@ -46,8 +40,6 @@ import java.util.List;
  */
 public class ModEventHandler {
     private final static String TAG = "EventHandler";
-    private final static ResourceLocation INVENTORY_SKILLS = new ResourceLocation("vampirism", "textures/gui/inventory_skills.png");
-    private final static int SKILLBUTTONID = 27496;
 
     @SubscribeEvent
     public void onAttachCapabilitiesVillage(AttachCapabilitiesEvent<Village> event) {
@@ -157,33 +149,5 @@ public class ModEventHandler {
     @SubscribeEvent
     public void onWorldUnload(WorldEvent.Unload event) {
         VampirismAPI.getGarlicChunkHandler(event.getWorld()).clear();
-    }
-
-    @SideOnly(Side.CLIENT)
-    @SubscribeEvent
-    public void onInitGuiEventPost(GuiScreenEvent.InitGuiEvent.Post event) {
-        if (Configs.gui_skill_button_enable && event.getGui() instanceof GuiInventory && FactionPlayerHandler.get(event.getGui().mc.player).getCurrentFactionPlayer() != null) {
-            List<GuiButton> buttonList = event.getButtonList();
-            GuiButton button = new GuiButtonImage(SKILLBUTTONID, ((GuiInventory) event.getGui()).getGuiLeft() + 125, event.getGui().height / 2 - 22, 20, 18, 178, 0, 19, INVENTORY_SKILLS);
-            buttonList.add(button);
-            event.setButtonList(buttonList);
-        }
-    }
-
-    @SideOnly(Side.CLIENT)
-    @SubscribeEvent
-    public void onActionPerformedPre(GuiScreenEvent.ActionPerformedEvent.Post event) {
-        if (Configs.gui_skill_button_enable && event.getGui() instanceof GuiInventory) {
-            if (event.getButton().id == SKILLBUTTONID) {
-                event.getGui().mc.displayGuiScreen(new GuiSkills());
-            } else if (event.getButton().id == 10) {
-                for (GuiButton e : event.getButtonList()) {
-                    if (e.id == SKILLBUTTONID) {
-                        ((GuiButtonImage) e).setPosition(((GuiInventory) event.getGui()).getGuiLeft() + 125, event.getGui().height / 2 - 22);
-                        break;
-                    }
-                }
-            }
-        }
     }
 }
